Reorganized classes with mixins
