+++ spikeinterface/comparison/__init__.py
@@ -3,11 +3,11 @@ from .comparisontools import (count_matching_events, compute_agreement_score, co
                               do_score_labels, compare_spike_trains, do_confusion_matrix, do_count_score,
                               compute_performance,
                               do_count_event, make_match_count_matrix)
-from .symmetricsortingcomparison import compare_two_sorters, SymmetricSortingComparison
-from .groundtruthcomparison import compare_sorter_to_ground_truth, GroundTruthComparison
-from .multisortingcomparison import compare_multiple_sorters, MultiSortingComparison
-from .templatecomparison import compare_templates, TemplateComparison
-from .multitemplatecomparison import compare_multiple_templates, MultiTemplateComparison
+from .paircomparisons import (compare_two_sorters, SymmetricSortingComparison, 
+                              compare_sorter_to_ground_truth, GroundTruthComparison,
+                              compare_templates, TemplateComparison)
+from .multicomparisons import (compare_multiple_sorters, MultiSortingComparison,
+                               compare_multiple_templates, MultiTemplateComparison)
 from .collisioncomparison import CollisionGTComparison
 from .correlogramcomparison import CorrelogramGTComparison
 from .groundtruthstudy import GroundTruthStudy
