+++ spikeinterface/comparison/basecomparison.py
@@ -1,6 +1,10 @@
 import numpy as np
-from .comparisontools import (make_possible_match, make_best_match, make_hungarian_match,
-                              do_count_event, make_match_count_matrix, make_agreement_scores_from_count)
+import networkx as nx
+
+from copy import deepcopy
+from typing import OrderedDict
+
+from .comparisontools import (make_possible_match, make_best_match, make_hungarian_match)
 
 
 class BaseComparison:
@@ -8,36 +12,233 @@ class BaseComparison:
     Base class for all comparisons (SpikeTrain and Template)
     """
 
-    def __init__(self, name_list,
+    def __init__(self, object_list, name_list,
                  match_score=0.5, chance_score=0.1,
                  verbose=False):
+        self.object_list = object_list
         self.name_list = name_list
         self._verbose = verbose
         self.match_score = match_score
         self.chance_score = chance_score
+
+
+class BaseMultiComparison(BaseComparison):
+    def __init__(self, object_list, name_list,
+                 match_score=0.5, chance_score=0.1,
+                 verbose=False):
+        BaseComparison.__init__(self, object_list=object_list,
+                                name_list=name_list,
+                                match_score=match_score,
+                                chance_score=chance_score,
+                                verbose=verbose)
+        self.graph = None
+        self.subgraphs = None
+        self.clean_graph = None
+
+    def _compute_all(self):
+        self._do_comparison()
+        self._do_graph()
+        self._clean_graph()
+        self._do_agreement()
+
+    def _compare_ij(self, i, j):
+        raise NotImplementedError
+
+    def _populate_nodes(self):
+        raise NotImplementedError
+
+    @property
+    def units(self):
+        return deepcopy(self._new_units)
+
+    def compute_subgraphs(self):
+        """
+        Computes subgraphs of connected components.
+        Returns
+        -------
+        sg_object_names: list
+            List of sorter names for each node in the connected component subgraph
+        sg_units: list
+            List of unit ids for each node in the connected component subgraph
+        """
+        if self.clean_graph is not None:
+            g = self.clean_graph
+        else:
+            g = self.graph
+        subgraphs = (g.subgraph(c).copy() for c in nx.connected_components(g))
+        sg_object_names = []
+        sg_units = []
+        for sg in subgraphs:
+            object_names = []
+            unit_names = []
+            for node in sg.nodes:
+                object_names.append(node[0])
+                unit_names.append(node[1])
+            sg_object_names.append(object_names)
+            sg_units.append(unit_names)
+        return sg_object_names, sg_units
+
+    def _do_comparison(self, ):
+        # do pairwise matching
+        if self._verbose:
+            print('Multicomaprison step 1: pairwise comparison')
+
+        self.comparisons = []
+        for i in range(len(self.object_list)):
+            for j in range(i + 1, len(self.object_list)):
+                if self._verbose:
+                    if self.name_list is not None:
+                        name_i = self.name_list[i]
+                        name_j = self.name_list[j]
+                    else:
+                        name_i = "object i"
+                        name_j = "object j"
+                    print(f"  Comparing: {name_i} and {name_j}")
+                comp = self._compare_ij(i, j)
+                self.comparisons.append(comp)
+
+    def _do_graph(self):
+        if self._verbose:
+            print('Multicomparison step 2: make graph')
+
+        self.graph = nx.Graph()
+        # nodes
+        self._populate_nodes()
+
+        # edges
+        for comp in self.comparisons:
+            for u1 in comp.hungarian_match_12.index.values:
+                u2 = comp.hungarian_match_12[u1]
+                if u2 != -1:
+                    node1 = comp.name_list[0], u1
+                    node2 = comp.name_list[1], u2
+                    score = comp.agreement_scores.loc[u1, u2]
+                    self.graph.add_edge(node1, node2, weight=score)
+
+        # the graph is symmetrical
+        self.graph = self.graph.to_undirected()
+
+    def _clean_graph(self):
+        if self._verbose:
+            print('Multicomaprison step 3: clean graph')
+        clean_graph = self.graph.copy()
+        subgraphs = (clean_graph.subgraph(c).copy()
+                     for c in nx.connected_components(clean_graph))
+        removed_nodes = 0
+        for sg in subgraphs:
+            object_names = []
+            for node in sg.nodes:
+                object_names.append(node[0])
+            sorters, counts = np.unique(object_names, return_counts=True)
+
+            if np.any(counts > 1):
+                for sorter in sorters[counts > 1]:
+                    nodes_duplicate = [n for n in sg.nodes if sorter in n]
+                    # get edges
+                    edges_duplicates = []
+                    weights_duplicates = []
+                    for n in nodes_duplicate:
+                        edges = sg.edges(n, data=True)
+                        for e in edges:
+                            edges_duplicates.append(e)
+                            weights_duplicates.append(e[2]['weight'])
+                    # remove edges
+                    edges_to_remove = len(nodes_duplicate) - 1
+                    remove_idxs = np.argsort(weights_duplicates)[
+                        :edges_to_remove]
+                    for idx in remove_idxs:
+                        if self._verbose:
+                            print('Removed edge', edges_duplicates[idx])
+                        clean_graph.remove_edge(
+                            edges_duplicates[idx][0], edges_duplicates[idx][1])
+                        sg.remove_edge(
+                            edges_duplicates[idx][0], edges_duplicates[idx][1])
+                        if edges_duplicates[idx][0] in nodes_duplicate:
+                            sg.remove_node(edges_duplicates[idx][0])
+                        else:
+                            sg.remove_node(edges_duplicates[idx][1])
+                        removed_nodes += 1
+        if self._verbose:
+            print(f'Removed {removed_nodes} duplicate nodes')
+        self.clean_graph = clean_graph
+
+    def _do_agreement(self):
+        # extract agreement from graph
+        if self._verbose:
+            print('Multicomparison step 4: extract agreement from graph')
+
+        self._new_units = {}
+
+        # save new units
+        self.subgraphs = (self.clean_graph.subgraph(c).copy()
+                          for c in nx.connected_components(self.clean_graph))
+        for new_unit, sg in enumerate(self.subgraphs):
+            edges = list(sg.edges(data=True))
+            if len(edges) > 0:
+                avg_agr = np.mean([d['weight'] for u, v, d in edges])
+            else:
+                avg_agr = 0
+            object_unit_ids = {}
+            for node in sg.nodes:
+                object_name, unit_name = node
+                object_unit_ids[object_name] = unit_name
+            # sort dict based on name list
+            sorted_object_unit_ids = OrderedDict()
+            for name in self.name_list:
+                if name in object_unit_ids:
+                    sorted_object_unit_ids[name] = object_unit_ids[name]
+            self._new_units[new_unit] = {'avg_agreement': avg_agr, 'unit_ids': sorted_object_unit_ids,
+                                         'agreement_number': len(sg.nodes)}
+
+
+class BasePairComparison(BaseComparison):
+    def __init__(self, object1, object2, name1, name2, 
+                 match_score=0.5, chance_score=0.1,
+                 verbose=False):
+        BaseComparison.__init__(self, object_list=[object1, object2],
+                                name_list=[name1, name2],
+                                match_score=match_score,
+                                chance_score=chance_score,
+                                verbose=verbose)
         self.possible_match_12, self.possible_match_21 = None, None
         self.best_match_12, self.best_match_21 = None, None
         self.hungarian_match_12, self.hungarian_match_21 = None, None
-
+        self.agreement_scores = None
+    
     def _do_agreement(self):
         # populate self.agreement_scores
-        NotImplementedError
+        raise NotImplementedError
 
     def _do_matching(self):
         if self._verbose:
             print("Matching...")
 
-        self.possible_match_12, self.possible_match_21 = make_possible_match(self.agreement_scores, self.chance_score)
-        self.best_match_12, self.best_match_21 = make_best_match(self.agreement_scores, self.chance_score)
-        self.hungarian_match_12, self.hungarian_match_21 = make_hungarian_match(self.agreement_scores, self.match_score)
+        self.possible_match_12, self.possible_match_21 = make_possible_match(
+            self.agreement_scores, self.chance_score)
+        self.best_match_12, self.best_match_21 = make_best_match(
+            self.agreement_scores, self.chance_score)
+        self.hungarian_match_12, self.hungarian_match_21 = make_hungarian_match(
+            self.agreement_scores, self.match_score)
+        
+    def get_ordered_agreement_scores(self):
+        assert self.agreement_scores is not None, "'agreement_scores' have not been computed!"
+        # order rows
+        order0 = self.agreement_scores.max(axis=1).argsort()
+        scores = self.agreement_scores.iloc[order0.values[::-1], :]
 
+        # order columns
+        indexes = np.arange(scores.shape[1])
+        order1 = []
+        for r in range(scores.shape[0]):
+            possible = indexes[~np.in1d(indexes, order1)]
+            if possible.size > 0:
+                ind = np.argmax(scores.iloc[r, possible].values)
+                order1.append(possible[ind])
+        remain = indexes[~np.in1d(indexes, order1)]
+        order1.extend(remain)
+        scores = scores.iloc[:, order1]
 
-class BaseMultiComparison(BaseComparison):
-    pass
-
-
-class BasePairComparison(BaseComparison):
-    pass
+        return scores
 
 
 class MixinSpikeTrainComparison:
@@ -53,40 +254,34 @@ class MixinSpikeTrainComparison:
       * delta_time to delta_frames
     """
     def __init__(self, delta_time=0.4, n_jobs=-1):
-        # self.sorting_list = sorting_list
-        # if name_list is None:
-        #     name_list = ['sorting{}'.format(i + 1)
-        #                  for i in range(len(sorting_list))]
-        # BaseComparison.__init__(self, name_list=name_list,
-        #                         match_score=match_score, chance_score=chance_score,
-        #                         verbose=verbose)
         self.delta_time = delta_time
         self.n_jobs = n_jobs
+        self.sampling_frequency = None
+        self.delta_frames = None
+        
+    def set_frames_and_frequency(self, sorting_list):
+        if not np.all(sorting.get_num_segments() == 1 for sorting in sorting_list):
+            raise Exception(
+                'Comparison module work with sorting having num_segments=1')
 
-        # self.name_list = name_list
-        # if np.any(['_' in name for name in name_list]):
-        #     raise ValueError("Sorter names in 'name_list' cannot contain '_'")
+        # take sampling frequency from sorting list and test that they are equivalent.
+        sampling_freqs = np.array([sorting.get_sampling_frequency()
+                                  for sorting in sorting_list], dtype='float64')
 
-        # if not np.all(s.get_num_segments() == 1 for s in sorting_list):
-        #     raise Exception('Comparison module work with sorting having num_segments=1')
+        # Some sorter round the sampling freq lets emit a warning
+        sf0 = sampling_freqs[0]
+        if not np.all(sf0 == sampling_freqs):
+            delta_freq_ratio = np.abs(sampling_freqs - sf0) / sf0
+            # tolerance of 0.1%
+            assert np.all(
+                delta_freq_ratio < 0.001), "Inconsistent sampling frequency among sorting list"
 
-        # # take sampling frequency from sorting list and test that they are equivalent.
-        # sampling_freqs = np.array([s.get_sampling_frequency() for s in self.sorting_list], dtype='float64')
-
-        # # Some sorter round the sampling freq lets emit a warning
-        # sf0 = sampling_freqs[0]
-        # if not np.all(sf0 == sampling_freqs):
-        #     delta_freq_ratio = np.abs(sampling_freqs - sf0) / sf0
-        #     # tolerance of 0.1%
-        #     assert np.all(delta_freq_ratio < 0.001), "Inconsistent sampling frequency among sorting list"
-
-        # self.sampling_frequency = sf0
-        # self.delta_time = delta_time
-        # self.delta_frames = int(self.delta_time / 1000 * self.sampling_frequency)
-        # self._n_jobs = n_jobs
+        self.sampling_frequency = sf0
+        self.delta_frames = int(
+            self.delta_time / 1000 * self.sampling_frequency)
 
 
 class MixinTemplateComparison:
-    pass
-
-
+    def __init__(self, similarity_method="cosine_similarity", sparsity_dict=None):
+        self.similarity_method = similarity_method
+        self.sparsity_dict = sparsity_dict
